+++ src/event_handler.rs
@@ -19,156 +19,10 @@ use libc::{O_RDWR, O_WRONLY};
 use miette::{miette, Result};
 use nix::poll::{poll, PollFd, PollFlags};
 // use serde::{Deserialize, Serialize};
-use knuffel::{Decode, DecodeScalar};
 
 use crate::config::Config;
+use crate::gestures::{hold::*, pinch::*, swipe::*, *};
 use crate::utils::exec_command_from_string;
-
-/// Direction of swipe gestures
-///
-/// NW  N  NE
-/// W   C   E
-/// SW  S  SE
-#[derive(DecodeScalar, Debug, Clone, PartialEq, Eq)]
-pub enum SwipeDir {
-    Any,
-    N,
-    S,
-    E,
-    W,
-    NE,
-    NW,
-    SE,
-    SW,
-}
-
-impl SwipeDir {
-    // This code is sort of a mess
-    pub fn dir(x: f64, y: f64) -> SwipeDir {
-        if x.abs() == 0.0 && y.abs() == 0.0 {
-            return SwipeDir::Any;
-        }
-        let oblique_ratio = 0.414;
-        if x.abs() > y.abs() {
-            let sd = if x < 0.0 { SwipeDir::W } else { SwipeDir::E };
-            if y.abs() / x.abs() > oblique_ratio {
-                if sd == SwipeDir::W {
-                    if y < 0.0 {
-                        SwipeDir::NW
-                    } else {
-                        SwipeDir::SW
-                    }
-                } else if sd == SwipeDir::E {
-                    if y < 0.0 {
-                        SwipeDir::NE
-                    } else {
-                        SwipeDir::SE
-                    }
-                } else {
-                    SwipeDir::Any
-                }
-            } else {
-                sd
-            }
-        } else {
-            let sd = if y < 0.0 { SwipeDir::N } else { SwipeDir::S };
-            if x.abs() / y.abs() > oblique_ratio {
-                if sd == SwipeDir::N {
-                    if x < 0.0 {
-                        SwipeDir::NW
-                    } else {
-                        SwipeDir::NE
-                    }
-                } else if sd == SwipeDir::S {
-                    if x < 0.0 {
-                        SwipeDir::SW
-                    } else {
-                        SwipeDir::SE
-                    }
-                } else {
-                    SwipeDir::Any
-                }
-            } else {
-                sd
-            }
-        }
-    }
-}
-
-/// Direction of pinch gestures
-#[derive(DecodeScalar, Debug, Clone, PartialEq, Eq)]
-pub enum PinchDir {
-    In,
-    Out,
-    Clockwise,
-    CounterClockwise,
-    Any,
-}
-
-impl PinchDir {
-    pub fn dir(scale: f64, delta_angle: f64) -> Self {
-        // If the scale is low enough, see if there is any rotation
-        // These values seem to work fairly well overall
-        // But maybe could be improved by checking here for large rotation
-        if (scale > 0.92) && (scale < 1.08) {
-            if delta_angle > 0.0 {
-                Self::Clockwise
-            } else {
-                Self::CounterClockwise
-            }
-        // Otherwise we have a normal pinch
-        } else if scale > 1.0 {
-            Self::Out
-        } else {
-            Self::In
-        }
-    }
-}
-
-#[derive(Decode, Debug, Clone, PartialEq)]
-pub enum Gesture {
-    Swipe(Swipe),
-    Pinch(Pinch),
-    Hold(Hold),
-    // Rotate(Rotate),
-    None,
-}
-
-#[derive(Decode, Debug, Clone, PartialEq, Eq)]
-pub struct Swipe {
-    #[knuffel(property)]
-    pub direction: SwipeDir,
-    #[knuffel(property)]
-    pub fingers: i32,
-    #[knuffel(property)]
-    pub update: Option<String>,
-    #[knuffel(property)]
-    pub start: Option<String>,
-    #[knuffel(property)]
-    pub end: Option<String>,
-}
-
-#[derive(Decode, Debug, Clone, PartialEq, Eq)]
-pub struct Pinch {
-    #[knuffel(property)]
-    pub fingers: i32,
-    #[knuffel(property)]
-    pub direction: PinchDir,
-    #[knuffel(property)]
-    pub update: Option<String>,
-    #[knuffel(property)]
-    pub start: Option<String>,
-    #[knuffel(property)]
-    pub end: Option<String>,
-}
-
-#[derive(Decode, Debug, Clone, PartialEq, Eq)]
-pub struct Hold {
-    #[knuffel(property)]
-    pub fingers: i32,
-    #[knuffel(property)]
-    pub action: Option<String>,
-}
 
 #[derive(Debug)]
 pub struct EventHandler {
